Added BoundChecker

+++ spacearena-core/src/main/java/ru/spacearena/engine/graphics/Image.java
@@ -6,7 +6,7 @@ package ru.spacearena.engine.graphics;
  */
 public interface Image {
 
-    float getWidth();
-    float getHeight();
+    int getWidth();
+    int getHeight();
 
 }
